list_hdf() & read_dict_from_hdf(): Add glob-style pattern matching
When the pattern parameter is set to a glob-style pattern list_hdf() only returns the names of the nodes and groups in an HDF5 file which match the pattern. In the same way read_dict_from_hdf() only returns the nodes which match the pattern, when the pattern parameter is specified. This should simplify the implementation of lazy loading approaches.